requirements.txt
matplotlib==2.1.1
mysql-connector-python==8.0.5
nltk==3.2.5
numpy==1.13.3
oauthlib (2.0.7)
pandas==0.22.0
pip==9.0.1
textblob==0.15.1
tweepy==3.6.0
urllib3==1.22
